requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==9.0.1
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.1.0
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
cssselect2==0.7.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.2
django-allauth==0.52.0
django-channels==0.7.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.8.0
frozenlist==1.3.3
graphene==3.2.1
graphene-django==2.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
html2pdf==0.1
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iso8601==1.1.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.1
multidict==6.0.4
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.23.4
oauthlib==3.2.2
openai==0.26.1
opencv-python==4.7.0.68
oscrypto==1.3.0
packaging==21.3
pandas==1.5.1
parsel==1.7.0
Pillow==9.2.0
pipenv==2022.11.25
platformdirs==2.5.4
promise==2.3
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyPDF3==1.0.6
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
queuelib==1.6.2
razorpay==1.3.0
reportlab==3.6.12
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
Scrapy==2.7.1
service-identity==21.1.0
singledispatch==4.0.0
six==1.16.0
sqlclient==21.4.24
sqlparse==0.4.3
stripe==5.0.0
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.2.1
tldextract==3.4.0
tqdm==4.64.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.13
virtualenv==20.17.0
virtualenv-clone==0.5.7
w3lib==2.1.0
webencodings==0.5.1
whitenoise==6.3.0
xhtml2pdf==0.2.8
yarl==1.8.2
zope.interface==5.5.2
